Use fixed-size types for JSON_INT_VALUE and JSON_BOOL_VALUE in ovis_json
Using 'int' type in varargs can cause type mismatches when call sites pass
different sized values, leading to value truncation or stack corruption.

Change va_arg() from int to int64_t for JSON_INT_VALUE and int32_t for
JSON_BOOL_VALUE. Update call sites of json_entity_new() and json_dict_build()
to pass appropriately sized values and add documentation clarifying the type
requirements.

Fixes issue #1626

+++ ldms/src/ldmsd/ldmsd_request.c
@@ -9368,7 +9368,7 @@ static json_entity_t __ldmsd_stat2dict(struct ldmsd_stat *stat)
 				JSON_FLOAT_VALUE, "max", stat->max,
 				JSON_FLOAT_VALUE, "max_ts", max_ts,
 				JSON_FLOAT_VALUE, "avg", stat->avg,
-				JSON_INT_VALUE, "count", stat->count,
+				JSON_INT_VALUE, "count", (int64_t)stat->count,
 				JSON_FLOAT_VALUE, "start_ts", start_ts,
 				JSON_FLOAT_VALUE, "end_ts", end_ts,
 				-1);
